+++ atlasdb-client/src/main/java/com/palantir/atlasdb/util/SlidingWindowMetricsInvocationHandler.java
@@ -1,0 +1,100 @@
+/*
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.atlasdb.util;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import java.lang.reflect.Method;
+import java.util.concurrent.TimeUnit;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.SlidingTimeWindowArrayReservoir;
+import com.codahale.metrics.Timer;
+import com.palantir.logsafe.UnsafeArg;
+import com.palantir.tritium.event.AbstractInvocationEventHandler;
+import com.palantir.tritium.event.DefaultInvocationContext;
+import com.palantir.tritium.event.InstrumentationProperties;
+import com.palantir.tritium.event.InvocationContext;
+
+/**
+ * A modified version of {@link com.palantir.tritium.event.metrics.MetricsInvocationEventHandler} that uses a
+ * {@link SlidingTimeWindowArrayReservoir} for the timer's reservoir.
+ */
+public final class SlidingWindowMetricsInvocationHandler extends AbstractInvocationEventHandler<InvocationContext> {
+    private static final Logger logger = LoggerFactory.getLogger(SlidingWindowMetricsInvocationHandler.class);
+    private static final String FAILURES_METRIC_NAME = "failures";
+
+    private final MetricRegistry metricRegistry;
+    private final String serviceName;
+
+    public SlidingWindowMetricsInvocationHandler(MetricRegistry metricRegistry, String serviceName) {
+        super(InstrumentationProperties.getSystemPropertySupplier(serviceName));
+        this.metricRegistry = checkNotNull(metricRegistry, "metricRegistry");
+        this.serviceName = checkNotNull(serviceName, "serviceName");
+    }
+
+    @Override
+    public InvocationContext preInvocation(Object instance, Method method, Object[] args) {
+        return DefaultInvocationContext.of(instance, method, args);
+    }
+
+    @Override
+    public void onSuccess(@Nullable InvocationContext context, @Nullable Object result) {
+        if (context == null) {
+            logger.debug("Encountered null metric context likely due to exception in preInvocation");
+            return;
+        }
+
+        long nanos = System.nanoTime() - context.getStartTimeNanos();
+        metricRegistry.timer(getBaseMetricName(context), SlidingWindowMetricsInvocationHandler::createNewTimer)
+                .update(nanos, TimeUnit.NANOSECONDS);
+    }
+
+    @Override
+    public void onFailure(@Nullable InvocationContext context, @Nonnull Throwable cause) {
+        if (context == null) {
+            markGlobalFailure();
+            logger.debug("Encountered null metric context likely due to exception in preInvocation: {}",
+                    UnsafeArg.of("cause", cause),
+                    cause);
+            return;
+        }
+
+        markGlobalFailure();
+        String failuresMetricName = MetricRegistry.name(getBaseMetricName(context), FAILURES_METRIC_NAME);
+        metricRegistry.meter(failuresMetricName).mark();
+        metricRegistry.meter(MetricRegistry.name(failuresMetricName, cause.getClass().getName())).mark();
+    }
+
+    private static Timer createNewTimer() {
+        return new Timer(new SlidingTimeWindowArrayReservoir(35, TimeUnit.SECONDS));
+    }
+
+    private void markGlobalFailure() {
+        metricRegistry.meter(FAILURES_METRIC_NAME).mark();
+    }
+
+    private String getBaseMetricName(InvocationContext context) {
+        return MetricRegistry.name(serviceName, context.getMethod().getName());
+    }
+}