+++ docs/source/release_notes/release-notes.rst
@@ -50,8 +50,10 @@ develop
     *    - Type
          - Change
 
-    *    -
-         -
+    *    - |metrics| |changed|
+         - All non-tagged AtlasDB metrics now use a ``SlidingTimeWindowArrayReservoir``.
+           Previously, they used an exponentially decaying reservoir.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/????>`__)
 
 ========
 v0.133.0